experiments/requirements.txt
openai>=1.0.0
retry
tqdm
pyyaml